switch to gettimeofday() for getting the current time
i should have known better in the first place:

http://stackoverflow.com/a/23597725/98653 (my answer) and the code i used in
fluxbox to correctly show the time use gettimeofday() for better accuracy.
using time() cuts off the fraction part of the struct that is used in kernel
land to hold the time. the result is that the rendered time looks ~0.5s behind
other, more accurat clocks.